Refactor movement code
Put everything that was in Run() inside of ProcessMovementAndGetVector() to reduce needless abstraction

Simplify jerk development by removing rate in favour of only having a jerk period variable

Tried to add clarity to the two different thrust variables with comments and moving them onto the same line. One is thrust which is multiplied by alignment jerk is added to. The other is a multiplier after all of that.

Removed unused exponential decay method